% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate_mattermost}
\alias{authenticate_mattermost}
\title{Authenticate with Mattermost API}
\usage{
authenticate_mattermost(
  base_url,
  token = NULL,
  username = NULL,
  password = NULL,
  test_connection = FALSE
)
}
\arguments{
\item{base_url}{The base URL of the Mattermost server. Do not add the team name here so really only  "https://yourmattermost.stackhero-network.com/"}

\item{token}{Optional. The Bearer token for authentication. If NULL, it will be retrieved from options().}

\item{username}{Optional. The username for login if not using a token.}

\item{password}{Optional. The password for login if not using a token.}

\item{test_connection}{Boolean. If `TRUE`, the function will check the connection status with Mattermost.}
}
\value{
A `mattermost_auth` object containing `base_url` and `headers` for further API calls.
}
\description{
Authenticate with Mattermost API
}
\examples{
\dontrun{
authenticate_mattermost(base_url = "https://mattermost.stackhero-network.com"
, token = "your token", test_connection = TRUE)
}
}
