% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{check_mattermost_auth}
\alias{check_mattermost_auth}
\title{Check if the object is a valid mattermost_auth object}
\usage{
check_mattermost_auth(auth)
}
\arguments{
\item{auth}{The object to check.}
}
\value{
NULL if the object is valid, otherwise throws an error.
}
\description{
Check if the object is a valid mattermost_auth object
}
