% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_channel.R
\name{create_channel}
\alias{create_channel}
\title{Create a new Mattermost channel}
\usage{
create_channel(
  team_id = NULL,
  name = NULL,
  display_name = NULL,
  type = "O",
  verbose = FALSE,
  auth = get_default_auth()
)
}
\arguments{
\item{team_id}{The ID of the team where the channel will be created.}

\item{name}{A short name for the channel.}

\item{display_name}{The display name for the channel.}

\item{type}{The type of the channel: "O" for open (default) or "P" for private.}

\item{verbose}{(Logical) If `TRUE`, detailed information about the request and response will be printed.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
A named list with the created channel's fields (e.g. `id`,
  `display_name`, `name`, `type`, `team_id`).
}
\description{
Create a new Mattermost channel
}
\examples{
\dontrun{
teams <- get_all_teams()
new_channel <- create_channel(team_id = teams$id[1], name = "newchannel2"
, display_name = "a new channel", verbose = TRUE)
}
}
