% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_channel.R
\name{delete_channel}
\alias{delete_channel}
\title{Delete a Mattermost channel}
\usage{
delete_channel(
  channel_id,
  team_id,
  verbose = FALSE,
  auth = authenticate_mattermost()
)
}
\arguments{
\item{channel_id}{The ID of the channel that will be deleted.}

\item{team_id}{The ID of the team to which the channel belongs.}

\item{verbose}{(Logical) If `TRUE`, detailed information about the request and response will be printed.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
A list with details about the deletion status.
}
\description{
Delete a Mattermost channel
}
\examples{
# First create a channel
teams <- get_all_teams()
new_channel <- create_channel(team_id = teams$id, name = "newchannel"
, display_name = "a new channel")

# Now delete the channel
info <- delete_channel(channel_id = new_channel$id, team_id = teams$id)
}
