% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_old_messages.R
\name{delete_old_messages}
\alias{delete_old_messages}
\title{Delete Messages Older Than X Days in a Mattermost Channel}
\usage{
delete_old_messages(channel_id, days, auth = get_default_auth())
}
\arguments{
\item{channel_id}{Character string. The ID of the Mattermost channel from which
messages should be deleted.}

\item{days}{Numeric. The age in days of the messages to be deleted.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
A data frame with columns `message_id` and `delete_status`, or an
  empty data frame if no messages were deleted.
}
\description{
This function deletes messages that are older than a specified number of days
in a given Mattermost channel. It auto-paginates through all posts in the
channel, filters by creation timestamp, and deletes the ones older than the
specified number of days.
}
\examples{
\dontrun{
  teams <- get_all_teams()
  team_channels <- get_team_channels(team_id = teams$id[1])
  channel_id <- get_channel_id_lookup(team_channels, name = "off-topic")
  posts <- delete_old_messages(channel_id, 0)

}
}
