% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bots.R
\name{disable_bot}
\alias{disable_bot}
\title{Disable a bot account}
\usage{
disable_bot(bot_user_id, verbose = FALSE, auth = get_default_auth())
}
\arguments{
\item{bot_user_id}{Character. The bot's user ID.}

\item{verbose}{Logical. Print detailed request information? Default `FALSE`.}

\item{auth}{A `mattermost_auth` object created by
[authenticate_mattermost()]. Defaults to [get_default_auth()].}
}
\value{
A named list representing the disabled bot (same schema as
  [create_bot()]).
}
\description{
Disable (soft-delete) a bot. The bot will no longer be able to post or
respond, but can be re-enabled with [enable_bot()].
}
\section{Permissions}{

Must have `manage_bots` permission for bots you own, or
`manage_others_bots` for bots owned by other users.
}

\examples{
\dontrun{
  disable_bot("bot_user_id_abc")
}
}
\seealso{
[enable_bot()], [get_bot()]
}
