% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_teams.R
\name{get_all_teams}
\alias{get_all_teams}
\title{List all teams in Mattermost}
\usage{
get_all_teams(verbose = FALSE, auth = authenticate_mattermost())
}
\arguments{
\item{verbose}{(Logical) If `TRUE`, detailed information about the request and response will be printed.}

\item{auth}{A list containing `base_url` and `headers` for authentication.}
}
\value{
A data frame containing details of all teams.
}
\description{
List all teams in Mattermost
}
\examples{
\dontrun{
teams <- get_all_teams()
}
}
