% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channel_info.R
\name{get_channel_info}
\alias{get_channel_info}
\title{Get information about a Mattermost channel}
\usage{
get_channel_info(channel_id, verbose = FALSE, auth = get_default_auth())
}
\arguments{
\item{channel_id}{A character string containing the Mattermost channel ID.}

\item{verbose}{(Logical) If `TRUE`, detailed information about the request
and response will be printed. Default is `FALSE`.}

\item{auth}{The authentication object created by [authenticate_mattermost()].}
}
\value{
A named list with channel fields returned by the Mattermost API
  (e.g. `id`, `display_name`, `name`, `type`, `team_id`).
}
\description{
Get information about a Mattermost channel
}
\examples{
\dontrun{
 get_channel_info(channel_id = "newchannel2", verbose = TRUE)
}
}
