% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channel_posts.R
\name{get_channel_posts}
\alias{get_channel_posts}
\title{Get posts from a Mattermost channel}
\usage{
get_channel_posts(
  channel_id,
  verbose = FALSE,
  auth = authenticate_mattermost()
)
}
\arguments{
\item{channel_id}{The Mattermost channel ID.}

\item{verbose}{Boolean. If `TRUE`, the function will print request/response details for more information.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
Parsed JSON response with posts from the channel.
}
\description{
Get posts from a Mattermost channel
}
\examples{
\dontrun{
  teams <- get_all_teams()
  team_channels <- get_team_channels(team_id = teams$id[1])
  channel_id <- get_channel_id_lookup(team_channels, name = "off-topic")
  posts <- get_channel_posts(channel_id)
}
}
