% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{get_command}
\alias{get_command}
\title{Get a slash command}
\usage{
get_command(command_id, verbose = FALSE, auth = get_default_auth())
}
\arguments{
\item{command_id}{Character. The command ID.}

\item{verbose}{Logical. Print detailed request information? Default `FALSE`.}

\item{auth}{A `mattermost_auth` object created by
[authenticate_mattermost()]. Defaults to [get_default_auth()].}
}
\value{
A named list representing the command (same schema as
  [create_command()]).
}
\description{
Retrieve a command definition by its ID.
}
\section{Permissions}{

Must have `manage_slash_commands` permission for the team the command
belongs to.
}

\examples{
\dontrun{
  cmd <- get_command("cmd_id_abc")
  cmd$trigger
}
}
\seealso{
[list_commands()], [create_command()]
}
