% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mattermost_file.R
\name{get_mattermost_file}
\alias{get_mattermost_file}
\title{Retrieve a file from Mattermost}
\usage{
get_mattermost_file(file_id, auth = authenticate_mattermost())
}
\arguments{
\item{file_id}{The ID of the file to be retrieved.}

\item{auth}{A list containing `base_url` and `headers` for authentication.}
}
\value{
The response from the Mattermost API, which may contain the file data.
}
\description{
This function retrieves a file from Mattermost based on its file ID.
}
\examples{
\dontrun{
# Assuming you have already authenticated and obtained the file_id
file_id <- "your_file_id_here"
file_response <- get_mattermost_file(file_id = file_id)
print(file_response)

file_response <- get_mattermost_file(file_id = "i5rb43jei787jxcud7ekyyyyhc")
}
}
