% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_me}
\alias{get_me}
\title{Get information about which user is belonging to bearer key,}
\usage{
get_me(verbose = FALSE, auth = get_default_auth())
}
\arguments{
\item{verbose}{(Logical) If `TRUE`, detailed information about the request and response will be printed.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
a list with user information for the authentication key
}
\description{
Get information about which user is belonging to bearer key,
}
\examples{
\dontrun{
 get_me()
}
}
