% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactions.R
\name{get_reactions}
\alias{get_reactions}
\title{Get All Reactions on a Mattermost Post}
\usage{
get_reactions(post_id, verbose = FALSE, auth = get_default_auth())
}
\arguments{
\item{post_id}{The ID of the post to get reactions for.}

\item{verbose}{(Logical) If \code{TRUE}, detailed information about the
request and response will be printed.}

\item{auth}{The authentication object created by
\code{\link{authenticate_mattermost}()}. Must be a valid
\code{mattermost_auth} object.}
}
\value{
A data frame of reactions with columns \code{user_id},
  \code{post_id}, \code{emoji_name}, and \code{create_at}. Returns an
  empty result if there are no reactions.
}
\description{
Retrieves all reactions made by all users on a given post.
}
\details{
The caller must have \code{read_channel} permission for the channel the
post is in.
}
\examples{
\dontrun{
reactions <- get_reactions(post_id = "post_abc123")
reactions
}
}
