% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_channels.R
\name{get_team_channels}
\alias{get_team_channels}
\title{Get the list of channels for a team}
\usage{
get_team_channels(team_id = NULL, verbose = FALSE, auth = get_default_auth())
}
\arguments{
\item{team_id}{A character string containing the Mattermost team ID.}

\item{verbose}{(Logical) If `TRUE`, detailed information about the request
and response will be printed. Default is `FALSE`.}

\item{auth}{The authentication object created by [authenticate_mattermost()].}
}
\value{
A data frame with one row per channel (columns include `id`,
  `display_name`, `name`, `type`, `team_id`, etc.).
}
\description{
Get the list of channels for a team
}
\examples{
\dontrun{
authenticate_mattermost()
teams <- get_all_teams()
team_channels <- get_team_channels(team_id = teams$id[1])
}
}
\seealso{
[get_channel_id_by_display_name()]
}
