% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_info.R
\name{get_user_info}
\alias{get_user_info}
\title{Get information about a specific Mattermost user}
\usage{
get_user_info(user_id = NULL, auth = authenticate_mattermost())
}
\arguments{
\item{user_id}{The Mattermost user ID.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
Parsed JSON response with user information.
}
\description{
Get information about a specific Mattermost user
}
\examples{
\dontrun{
users <- get_all_users()
userinfo <- lapply(users, get_user_info)
}
}
