% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mattermost_api_request.R
\name{handle_response_content}
\alias{handle_response_content}
\title{Handle the content of the response}
\usage{
handle_response_content(response, verbose)
}
\arguments{
\item{response}{The response object from the API request.}

\item{verbose}{Boolean. If `TRUE`, the function will print the response details for more information.}
}
\value{
The response object if the content type is JSON, or a warning if it's not.
}
\description{
Handle the content of the response
}
