% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_message.R
\name{normalize_priority}
\alias{normalize_priority}
\title{Normalize the priority input}
\usage{
normalize_priority(priority)
}
\arguments{
\item{priority}{A string representing the priority.}
}
\value{
A string with the corrected priority format.
}
\description{
This function converts various casing inputs for priority to the correct format.
Valid priorities match the Mattermost UI: Normal, Important, Urgent.
}
