% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_mattermost_file.R
\name{send_mattermost_file}
\alias{send_mattermost_file}
\title{Send a file to a Mattermost channel}
\usage{
send_mattermost_file(
  channel_id,
  file_path,
  comment = NULL,
  auth = authenticate_mattermost(),
  verbose = FALSE
)
}
\arguments{
\item{channel_id}{The ID of the channel to which the file should be sent.}

\item{file_path}{The path to the file to be sent.}

\item{comment}{A comment to accompany the file.}

\item{auth}{A list containing `base_url` and `headers` for authentication.}

\item{verbose}{(Logical) If `TRUE`, detailed information about the request and response will be printed.}
}
\value{
The response from the Mattermost API.
}
\description{
This function sends a file to a specified Mattermost channel along with an optional comment
}
\examples{
\dontrun{
# Create a sample text file
fileconn <- file("output.txt")
writeLines(c("Hello", "world"), fileconn)
close(fileconn)

teams <- get_all_teams()
team_channels <- get_team_channels(team_id = teams$id[1])
channel_id <- get_channel_id_lookup(team_channels, "Off-Topic")
response <- send_mattermost_file(channel_id = channel_id
,file_path = "output.txt",comment = "A simple text file.", verbose = TRUE)
print(response)
unlink(fileconn)
}
}
