% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_message.R
\name{send_mattermost_message}
\alias{send_mattermost_message}
\title{Send a text message to a Mattermost channel}
\usage{
send_mattermost_message(
  channel_id,
  message,
  priority = NULL,
  file_path = NULL,
  comment = NULL,
  verbose = FALSE,
  auth = authenticate_mattermost()
)
}
\arguments{
\item{channel_id}{The ID of the Mattermost channel.}

\item{message}{The message content.}

\item{priority}{A string specifying the priority of the message. Must be one of:
- "Normal" (default)
- "High"
- "Low"}

\item{file_path}{The path to the file to be sent.}

\item{comment}{A comment to accompany the attachment file.}

\item{verbose}{Boolean. If `TRUE`, the function will print request/response details for more information.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
Parsed response from the Mattermost server.
}
\description{
Send a text message to a Mattermost channel
}
\examples{
\dontrun{
# Define channel ID and message
authenticate_mattermost()
teams <- get_all_teams()
team_channels <- get_team_channels(team_id = teams$id[1])
channel_id <- get_channel_id_lookup(team_channels, "Off-Topic")

message <- paste("Hello, Mattermost! This is a test message at", Sys.time())

# Send the message with Normal priority
response <- send_mattermost_message(channel_id = channel_id
, message = message, priority = "Low", verbose = TRUE)


#send message with attachment file
fileconn <- file("output.txt")
writeLines(c("Hello", "world"), fileconn)
close(fileconn)

response <- send_mattermost_message(channel_id = channel_id
, message = message, file_path = "output.txt", verbose = TRUE, priority = "High")
unlink(fileconn)
}
}
