% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mattermost_status.R
\name{check_mattermost_status}
\alias{check_mattermost_status}
\title{Check if the Mattermost server is online}
\usage{
check_mattermost_status(verbose = FALSE, auth = authenticate_mattermost())
}
\arguments{
\item{verbose}{Boolean. If `TRUE`, the function will print request/response details for more information.}

\item{auth}{A list containing `base_url` and `headers` for authentication.}
}
\value{
TRUE if the server is online; FALSE otherwise.
}
\description{
Check if the Mattermost server is online
}
\examples{
\dontrun{
check_mattermost_status()
}

}
