% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_old_messages.R
\name{delete_old_messages}
\alias{delete_old_messages}
\title{Delete Messages Older Than X Days in a Mattermost Channel}
\usage{
delete_old_messages(channel_id, days, auth = authenticate_mattermost())
}
\arguments{
\item{channel_id}{Character string. The ID of the Mattermost channel from which
messages should be deleted.}

\item{days}{Numeric. The age in days of the messages to be deleted.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
A character vector of message IDs that were deleted.
}
\description{
This function deletes messages that are older than a specified number of days
in a given Mattermost channel. It fetches the messages from the channel,
checks their timestamps, and deletes the ones that are older than the specified days.
}
\examples{
\dontrun{
  teams <- get_all_teams()
  team_channels <- get_team_channels(team_id = teams$id[1])
  channel_id <- get_channel_id_lookup(team_channels, name = "off-topic")
  posts <- delete_old_messages(channel_id, 0)

}
}
