% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_post.R
\name{delete_post}
\alias{delete_post}
\title{Delete a specific post in Mattermost}
\usage{
delete_post(post_id = NULL, verbose = FALSE, auth = get_default_auth())
}
\arguments{
\item{post_id}{The ID of the post to delete.}

\item{verbose}{Boolean. If `TRUE`, the function will print request/response details for more information.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
A named list with the deletion status as returned by the
  Mattermost API.
}
\description{
Delete a specific post in Mattermost
}
\examples{

\dontrun{
 delete_post(post_id = "fake_id")
}
}
