% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_teams.R
\name{get_all_teams}
\alias{get_all_teams}
\title{List all teams in Mattermost}
\usage{
get_all_teams(verbose = FALSE, auth = get_default_auth())
}
\arguments{
\item{verbose}{(Logical) If `TRUE`, detailed information about the request
and response will be printed. Default is `FALSE`.}

\item{auth}{The authentication object created by [authenticate_mattermost()].}
}
\value{
A data frame with one row per team (columns include `id`,
  `display_name`, `name`, `type`, etc.). Returns an empty data frame /
  list if the user belongs to no teams (with a warning).
}
\description{
List all teams in Mattermost
}
\examples{
\dontrun{
teams <- get_all_teams()
}
}
