% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_all_users}
\alias{get_all_users}
\title{Get information about a specific Mattermost user}
\usage{
get_all_users(verbose = FALSE, auth = authenticate_mattermost())
}
\arguments{
\item{verbose}{(Logical) If `TRUE`, detailed information about the request and response will be printed.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
a vector user_ids
}
\description{
Get information about a specific Mattermost user
}
\examples{
\dontrun{
 get_all_users()
}

}
