% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channel_info.R
\name{get_channel_info}
\alias{get_channel_info}
\title{Get information about a Mattermost channel}
\usage{
get_channel_info(channel_id, verbose = FALSE, auth = authenticate_mattermost())
}
\arguments{
\item{channel_id}{The Mattermost channel ID.}

\item{verbose}{(Logical) If `TRUE`, detailed information about the request and response will be printed.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
Parsed JSON response with channel information.
}
\description{
Get information about a Mattermost channel
}
\examples{
 get_channel_info(channel_id = "newchannel2", verbose = TRUE)

}
