% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_teams.R
\name{get_team}
\alias{get_team}
\title{Get data of a single team from its team_id}
\usage{
get_team(team_id = NULL, verbose = FALSE, auth = get_default_auth())
}
\arguments{
\item{team_id}{A character string containing the Mattermost team ID.}

\item{verbose}{(Logical) If `TRUE`, detailed information about the request
and response will be printed. Default is `FALSE`.}

\item{auth}{The authentication object created by [authenticate_mattermost()].}
}
\value{
A named list with team fields returned by the Mattermost API
  (e.g. `id`, `display_name`, `name`, `type`).
}
\description{
Get data of a single team from its team_id
}
\examples{
\dontrun{
teams <- get_all_teams()
teaminfo <- lapply(teams$id, get_team)
}
}
