% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_user}
\alias{get_user}
\title{Get information about a specific Mattermost user}
\usage{
get_user(user_id = NULL, verbose = FALSE, auth = authenticate_mattermost())
}
\arguments{
\item{user_id}{The ID of the post to delete.}

\item{verbose}{(Logical) If `TRUE`, detailed information about the request and response will be printed.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
a vector user_ids
}
\description{
Get information about a specific Mattermost user
}
\examples{
\dontrun{
 users <- get_all_users()
 userinfo <- lapply(users, get_user)
}

}
