% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_message.R
\name{send_mattermost_message}
\alias{send_mattermost_message}
\title{Send a text message to a Mattermost channel, optionally with multiple attachments}
\usage{
send_mattermost_message(
  channel_id,
  message,
  priority = "Normal",
  file_path = NULL,
  comment = NULL,
  verbose = FALSE,
  auth = authenticate_mattermost()
)
}
\arguments{
\item{channel_id}{The ID of the Mattermost channel.}

\item{message}{The message content.}

\item{priority}{A string specifying the priority of the message. Must be one of:
- "Normal" (default)
- "High"
- "Low"}

\item{file_path}{A vector of file paths to be sent as attachments.}

\item{comment}{A comment to accompany the attachment files.}

\item{verbose}{Boolean. If `TRUE`, the function will print request/response details for more information.}

\item{auth}{The authentication object created by `authenticate_mattermost()`.}
}
\value{
Parsed response from the Mattermost server.
}
\description{
This function sends a text message to a Mattermost channel, optionally including one or more attachment files.
}
\examples{
\dontrun{
# Define channel ID and message
authenticate_mattermost()
teams <- get_all_teams()
team_channels <- get_team_channels(team_id = teams$id[1])
channel_id <- get_channel_id_lookup(team_channels, "Off-Topic")

message <- paste("Hello, Mattermost! This is a test message at", Sys.time())

# Send the message with a plot attachment
tmp_plot <- tempfile(fileext = ".png")
plot <- ggplot2::ggplot(cars, ggplot2::aes(x = speed, y = dist)) +
  ggplot2::geom_point()

ggplot2::ggsave(filename = tmp_plot, plot = plot)

response <- send_mattermost_message(
  channel_id = channel_id,
  message = message,
  priority = "High",
  file_path = tmp_plot,
  verbose = TRUE
)
print(response)


# Send message with a text file attachment
fileconn <- file("output.txt")
writeLines(c("Hello", "world"), fileconn)
close(fileconn)

response <- send_mattermost_message(
  channel_id = channel_id,
  message = message,
  file_path = c("output.txt", tmp_plot),
  verbose = TRUE,
  priority = "High"
)
print(response)
unlink("output.txt")
unlink(tmp_plot)
}
}
